--- SETUP_SUMMARY.md
@@ -1,119 +1,0 @@
-# OpenHands CLI Setup Summary
-
-## ✅ Completed Tasks
-
-### 1. Repository Cloning
-- Successfully cloned the `OpenHands-CLI` repository from GitHub using the GitHub API
-- Repository location: `/workspace/project/OpenHands-CLI`
-
-### 2. TUI Code Migration
-- Copied TUI-related files from `OpenHands/openhands/cli` to the new repository
-- Files migrated:
-  - `tui.py` - Main TUI functionality
-  - `pt_style.py` - Prompt Toolkit styling
-  - `utils.py` - Utility functions
-  - `settings.py` - Settings management
-  - `main.py` - Original main entry point
-  - `entry.py` - Entry point utilities
-
-### 3. Project Configuration
-- Created `pyproject.toml` with both modern Python project format and Poetry compatibility
-- Added agent-sdk dependency from the specified commit: `37749a54c18bc628b6bf7517ad96286cecdb8a67`
-- Configured build system with hatchling
-- Added development dependencies: pytest, black, isort, flake8, mypy
-
-### 4. UV Package Management Setup
-- Installed UV package manager
-- Successfully configured UV for dependency management
-- Created virtual environment and installed all dependencies
-- UV sync working properly with both main and dev dependencies
-
-## 📁 Project Structure
-
-```
-OpenHands-CLI/
-├── openhands_cli/           # Main package directory
-│   ├── __init__.py         # Package initialization
-│   ├── simple_main.py      # Simplified main entry point
-│   ├── tui.py             # Terminal User Interface
-│   ├── pt_style.py        # Prompt Toolkit styling
-│   ├── utils.py           # Utility functions
-│   ├── settings.py        # Settings management
-│   ├── main.py            # Original main entry point
-│   └── entry.py           # Entry point utilities
-
-├── pyproject.toml         # Project configuration
-├── Makefile              # Build automation
-├── README.md             # Documentation
-├── LICENSE               # MIT License
-└── uv.lock              # UV lock file
-```
-
-## 🚀 Usage
-
-### Running with UV
-```bash
-# Install dependencies
-uv sync
-
-# Run the CLI
-uv run openhands-cli
-```
-
-### Running with Poetry
-```bash
-# Install dependencies
-poetry install
-
-# Run the CLI
-poetry run openhands-cli
-```
-
-
-
-### Using Make Commands
-```bash
-make help          # Show available commands
-make install       # Install dependencies
-make install-dev   # Install with dev dependencies
-make run           # Run the CLI
-make clean         # Clean build artifacts
-```
-
-## 🔧 Package Management
-
-The project uses UV for fast, modern Python package management with Poetry compatibility.
-
-## 📦 Dependencies
-
-### Main Dependencies
-- `prompt-toolkit>=3.0.0` - Terminal UI framework
-- `openhands` - Agent SDK from specified commit
-
-### Development Dependencies
-- `pytest>=7.0.0` - Testing framework
-- `black>=23.0.0` - Code formatter
-- `isort>=5.0.0` - Import sorter
-- `flake8>=6.0.0` - Linter
-- `mypy>=1.0.0` - Type checker
-
-
-## 🎯 Current Status
-
-The CLI is successfully set up with:
-- ✅ Working package structure
-- ✅ UV package management
-- ✅ Poetry compatibility
-- ✅ Build automation with Make
-- ✅ Development tools configured
-
-The CLI currently runs a simplified version that demonstrates the setup. To enable full TUI functionality, additional dependencies and configuration may be needed based on the specific OpenHands agent requirements.
-
-## 🔄 Next Steps
-
-To extend functionality:
-1. Add missing dependencies (e.g., pydantic) as needed
-2. Configure agent SDK integration
-3. Set up environment variables and authentication
-4. Add comprehensive tests
-5. Enhance the TUI with full OpenHands integration
